+++ packages/img-cropper/index.js
@@ -1,0 +1,7 @@
+import ImgCropper from './src/main.vue'
+
+ImgCropper.install = Vue => {
+  Vue.component(ImgCropper.name, ImgCropper)
+}
+
+export default ImgCropper