+++ src/index.js
@@ -29,6 +29,7 @@ import Input from '../packages/input'
 import InputNumber from '../packages/input-number'
 import Img from '../packages/img'
 import ImgPreview from '../packages/img-preview'
+import ImgCropper from '../packages/img-cropper'
 import Lazyload from '../packages/lazyload'
 import Loading from '../packages/loading'
 import MessageBox from '../packages/message-box'
@@ -98,6 +99,7 @@ const components = [
   Input,
   InputNumber,
   Img,
+  ImgCropper,
   Loading,
   Modal,
   Navbar,
@@ -196,6 +198,7 @@ export default {
   InputNumber,
   Img,
   ImgPreview,
+  ImgCropper,
   Lazyload,
   Loading,
   MessageBox,